Add support for in process server/client
