+++ tests/src/test/resources/logback-test.xml
@@ -10,6 +10,7 @@
 
     <logger name="net.devh.springboot.autoconfigure.grpc.server" level="debug" />
     <logger name="net.devh.springboot.autoconfigure.grpc.server.security" level="debug" />
+    <logger name="net.devh.springboot.autoconfigure.grpc.client" level="debug" />
     <logger name="net.devh.test.grpc.server" level="debug" />
     <root level="info">
         <appender-ref ref="STDOUT" />
